long, Runtime,adminservice

+++ src/main/java/org/society/dao/UserDao.java
@@ -8,7 +8,7 @@ import org.society.exceptions.UserNotFoundException;
 public interface UserDao {
 	public int registerUser(User user);
 	public int updateUser(User user) throws UserNotFoundException;
-	public int deleteUser(int userId) throws UserNotFoundException;
+	public int deleteUser(long userId) throws UserNotFoundException;
 	public List<User> viewUserList();
-	public User findByUserId(int userId) throws UserNotFoundException;
+	public User findByUserId(long userId) throws UserNotFoundException;
 }
